started using jest and supertest and refactored code base to work with it
